LopShelve.jl Documentation

LopShelve.jl Documentation

What is LopShelve

Function

Index

What is LopShelve

A shelf is an object similar to a dictionary but which allows data persistence. You could thus save your data in a file as if it were a dictionary. This julia implementation is based on the Shelve version of python. So if you used the Shelve version of python you would not be out of place. However this implementation does not stop there, you could also use LopShelve to map your data from database (Sqlite) to dictionary

Function

Base.delete! — Method

delete(s::Shelf)

delete the shelve file and the shelf object

Base.in — Method

Base.in(item::Any, x::ShelfSql)

Return true if the item is in the table

julia> "a" in db => true
julia> ("a", "b") in db => true # multiple primary keys

Base.iterate — Function

Base.iterate(x::ShelfSql, state::Int64=1)::Union{Tuple{Any,Int64}, Nothing}

Iterate over the table and return the key and the value of each row. It's very useful to iterate through the whole table with a loop or check if an element exists in the table

for i in db
    print(i)
end

Base.keys — Method

Base.keys(x::ShelfSql)

Return the primary keys of the table

julia> keys(db) # Will return the primary keys of the table

Base.length — Method

Base.length(x::ShelfSql) :: Int64

Return the number of rows in the table

julia> length(db)  # 2 (We have 2 rows in the table)

Base.setindex! — Method

Base.setindex!(s::ShelfSql, value::NamedTuple, key)

Create a new row in the database or modify an existing row in the database Usually this method is called with bracket square in which we put the key's value and the row's value is the affected value.

    Exemple:
    db["justin"] = (year=2000, country="USA")
    # Multiple index values
    db[("John", "Doe")] = (year=1999, country="Benin")

LopShelve.open! — Method

open!(filename::String, table::String)

Open a given sqlite table from a given filename in a Shelf object

If the file don't exist an error is thrown

LopShelve.open! — Method

open(filename::String)

Open a given file in a Shelf object

If the file don't exist an empty Shelve object is created. Also the file name must'nt have extension

LopShelve.primary_key_finder — Method

primary_key_finder(db, table)

Find the primary of a given table.

LopShelve.save — Method

save(s::Shelf)

Save the shelve data

LopShelve.value_process — Method

value_process(key)

Process the key send in indexing by the user: if the key is a single value, we convert it to a tuple Moreover we add single quote to the string in order to bind the sql query

    value_process("a") => ("a", )
    value_process(("a", "b")) => ("a", "b")

Index

Base.delete!

Base.in

Base.iterate

Base.keys

Base.length

Base.setindex!

LopShelve.open!

LopShelve.open!

LopShelve.primary_key_finder

LopShelve.save

LopShelve.value_process